Lots of cleanups
