+++ src/agents/bg_task_agent/bg_task_agent.py
@@ -7,8 +7,7 @@ from langgraph.checkpoint.memory import MemorySaver
 from langgraph.graph import END, MessagesState, StateGraph
 
 from agents.bg_task_agent.task import Task
-from core import settings
-from core.llm import get_model
+from core import get_model, settings
 
 
 class AgentState(MessagesState, total=False):
