+++ src/agents/llama_guard.py
@@ -2,10 +2,10 @@ from enum import Enum
 
 from langchain_core.messages import AIMessage, AnyMessage, HumanMessage
 from langchain_core.prompts import PromptTemplate
-from langchain_groq import ChatGroq
 from pydantic import BaseModel, Field
 
-from core import settings
+from core import get_model, settings
+from schema.models import GroqModelName
 
 
 class SafetyAssessment(Enum):
@@ -81,9 +81,7 @@ class LlamaGuard:
             print("GROQ_API_KEY not set, skipping LlamaGuard")
             self.model = None
             return
-        self.model = ChatGroq(model="llama-guard-3-8b", temperature=0.0).with_config(
-            tags=["llama_guard"]
-        )
+        self.model = get_model(GroqModelName.LLAMA_GUARD_3_8B).with_config(tags=["llama_guard"])
         self.prompt = PromptTemplate.from_template(llama_guard_instructions)
 
     def _compile_prompt(self, role: str, messages: list[AnyMessage]) -> str:
