+++ src/core/__init__.py
@@ -1,3 +1,4 @@
-from .config import settings
+from core.llm import get_model
+from core.settings import settings
 
-__all__ = ["settings"]
+__all__ = ["settings", "get_model"]
