+++ src/schema/models.py
@@ -1,0 +1,47 @@
+from enum import StrEnum, auto
+from typing import TypeAlias
+
+
+class Provider(StrEnum):
+    OPENAI = auto()
+    ANTHROPIC = auto()
+    GOOGLE = auto()
+    GROQ = auto()
+    AWS = auto()
+
+
+class OpenAIModelName(StrEnum):
+    """https://platform.openai.com/docs/models/gpt-4o"""
+
+    GPT_4O_MINI = "gpt-4o-mini"
+    GPT_4O = "gpt-4o"
+
+
+class AnthropicModelName(StrEnum):
+    """https://docs.anthropic.com/en/docs/about-claude/models#model-names"""
+
+    HAIKU_3 = "claude-3-haiku"
+    HAIKU_35 = "claude-3.5-haiku"
+    SONNET_35 = "claude-3.5-sonnet"
+
+
+class GoogleModelName(StrEnum):
+    GEMINI_15_FLASH = "gemini-1.5-flash"
+
+
+class GroqModelName(StrEnum):
+    """https://console.groq.com/docs/models"""
+
+    LLAMA_31_8B = "groq-llama-3.1-8b"
+    LLAMA_31_70B = "groq-llama-3.1-70b"
+
+    LLAMA_GUARD_3_8B = "groq-llama-guard-3-8b"
+
+
+class AWSModelName(StrEnum):
+    BEDROCK_HAIKU = "bedrock-3.5-haiku"
+
+
+AllModelEnum: TypeAlias = (
+    OpenAIModelName | AnthropicModelName | GoogleModelName | GroqModelName | AWSModelName
+)
